Best word finder v3 time baybeee
Adds the 3rd version of the best word finder. this does a couple things but mainly it uses only the solution words mutiplying the min speed by about 10 as well as fixing an issue with double letters as well as fixing the dataframe printing. this is the first truly working version as of when i write this.