More File management and new files
This commit contains more file management as well as the addition of the average Score calculator, which calculates the average score of the current algorithm which is too slow, as well as changes to the bestWordFinder function. I beleieve that this is also the first commit with the bestWordFinder v3 the end of that algorithm for now, it also adds tkinterWordleClonePolishedV2 which  adds the feature of a hint button however it is way too slow and causes minute long game freeezes.